+++ src/model/port/ModelPortProtocol.ts
@@ -27,7 +27,7 @@ export interface PredictMessage extends PredictRequestBase<"predict"> {
     /** State data. */
     state: Float32Array[];
     /** Choice legality mask. */
-    choices: Uint8Array;
+    choices: Float32Array;
     /** Id of the previous action. Used for experience generation. */
     lastAction?: number;
     /** Reward from the state transition. Used for experience generation. */
